refactor: Rename /lights to /Light + descriptor typedefs
