+++ public/collision/Scene.js
@@ -1,5 +1,5 @@
 import {Mesh, Scene as _Scene} from "../../src/index.js";
-import {PointLight} from "../../src/lights/index.js";
+import {PointLight} from "../../src/Light/index.js";
 import {Material} from "../../src/materials/index.js";
 import {Matrix4, Vector3} from "../../src/math/index.js";
