+++ public/collision/scenes/swept_aabb.js
@@ -1,6 +1,6 @@
 import {Camera} from "../../../src/index.js";
 import {BoxGeometry} from "../../../src/Geometry/BoxGeometry.js";
-import {PointLight} from "../../../src/lights/index.js";
+import {PointLight} from "../../../src/Light/index.js";
 import {SSDLoader} from "../../../src/Loader/index.js";
 import {Material} from "../../../src/materials/Material.js";
 import {Matrix3, PI, Vector2, Vector3} from "../../../src/math/index.js";
