+++ public/hl2/scenes/building_entrance.js
@@ -1,5 +1,5 @@
 import {Camera} from "../../../src/index.js";
-import {PointLight} from "../../../src/lights/index.js";
+import {PointLight} from "../../../src/Light/index.js";
 import {SSDLoader} from "../../../src/Loader/index.js";
 import {PI, Vector2, Vector3} from "../../../src/math/index.js";
 import {ENTITY_HEIGHT_STAND, FIELD_OF_VIEW, SENSITIVITY} from "../main.js";
