+++ src/Light/AmbientLight.js
@@ -1,0 +1,3 @@
+import {Light} from "./Light.js";
+
+export class AmbientLight extends Light {}
