+++ src/Light/index.js
@@ -1,4 +1,4 @@
-export {AbstractLight} from "./AbstractLight.js";
 export {AmbientLight} from "./AmbientLight.js";
 export {DirectionalLight} from "./DirectionalLight.js";
+export {Light} from "./Light.js";
 export {PointLight} from "./PointLight.js";
