--- src/lights/AmbientLight.js
@@ -1,3 +1,0 @@
-import {AbstractLight} from "./AbstractLight.js";
-
-export class AmbientLight extends AbstractLight {}
